Split authors on "and" delimiter, creating and updating bibtex alist
BibTeX and BibLaTeX use "and" as the delimiter between author names. So, we
convert this single string of authors into a list of string based on the
location of the "and" in the string.

Properly remove and then add authors key to bibtex entry alist

Reverse the final author list when done

We read the author string from left to right, and cons each author onto the
front as we go, which means the final list is backwards from the order present
in the given string. Reverse it to correct this.