Tags have 2 forms, a braces form and a quoted form
The brace form allows for any text inside the braces, and braces inside the
braces should be escaped. The quote form allows for any text inside the
quotes, except quotes must be escaped.

The other key difference is that BibTeX allows string concatenation using the
\# character when both strings use quotes. See http://www.bibtex.org/Format/
for more information about this.