New test assertion macro for comparing alists
By default, equal? will perform a ordering-sensitive comparison, which usually
makes sense. However, for alists, this is not quite as nice, because the
ordering of elements within the alist is not important here. Rather we care
that the two alists have the same entries and that the entries have the same
values.