Upon starting gloa, start a REPL
This takes some inspiration from how Guix starts a REPL when using the
guix-repl command. In fact, when you invoke "guix repl", you call the
guix-repl command with a guile-type REPL with no specification. This is
because you are just running a Guile REPL inside your program.

If you want more advanced features (remote REPLs using sockets), that is NOT
supported with Gloa.