Add string-split-substring function for splitting strings
This behaves like Python's string-splitting mechanism. You provide a string to
split and the substring that should be used as the delimiter for splitting the
first string, returning a list of strings.

This has to be maintained separately until it gets merged into upstream Guile,
if it does.
From https://www.mail-archive.com/guile-devel@gnu.org/msg16045.html