Add ops module to pull together store & db operations to single command
It is easier (and more correct) to only expose the add-document procedure
rather than both the add-to-store and add-to-db procedures. Exposing those two
lower-level APIs allows us to reach an inconsistent state between the store
and the database.

In the long run, we should have a way to verify and perhaps correct either the
store or database when they become inconsistent with each other. This would
function similarly to Guix's "guix gc --verify" command.