Fix parallel testing
Introduce a function to produce somewhat random file paths for database
testing. This is built around Guile's tmpnam procedure, which does not
always produce a unique name. BUT, if this is run in Guix's build
environment, the /tmp in the build container is almost empty anyways.

When a batch of SRFI-64 tests are run through Make with the -jN flag
passed, where N > 1, each test-begin/test-end (which is a test suite)
runs IN PARALLEL!

This meant that the database tests which both use test.db as their test
database path would experience a race condition when operating on their
testing databases. Depending on which suite finished first, the test
database would be removed at the end of the suite, causing the other
suite to fail. Using somewhat random names for each will prevent this
kind of problem (unless the random names are not random enough).