Add utility function to delete a file recursively (including dirs)
This function behaves almost identically (at least I think so) to the common
shell command "rm -r file", where if file is a directory, the entire directory
is recursively emptied, then the directory that is file is also deleted.

This procedure is taken directly from Guix's (guix build utils), with minor
modifications to make reading it easier for me.