Add utility function to find a file's extension
This is useful when adding a document to the store. If the user provides a
document with an extension, then we can preserve that extension for their
other programs. However, if that file did not have an extension in its
original name, then we do not try to do anything.